code(): Extract the schema in own module
